Fix performance issues with getAllTags

+++ model/src/main/java/com/bookstore/dto/TagDto.java
@@ -2,20 +2,20 @@ package com.bookstore.dto;
 
 public class TagDto {
 
-    public TagDto(String name, Integer count) {
+    public TagDto(String name, Long count) {
         this.name = name;
         this.count = count;
     }
 
     private final String name;
 
-    private final Integer count;
+    private final Long count;
 
     public String getName() {
         return name;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 }
